+++ package.json
@@ -1,88 +1,88 @@
 {
-  "name": "duos-ui",
-  "version": "1.0.0",
-  "private": true,
-  "dependencies": {
-    "@databiosphere/bard-client": "^0.1.0",
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "11.11.0",
-    "@mui/icons-material": "5.14.1",
-    "@mui/material": "5.14.2",
-    "ajv": "8.12.0",
-    "ajv-formats": "2.1.1",
-    "axios": "^1.4.0",
-    "bootstrap": "3.4.1",
-    "dompurify": "3.0.5",
-    "http-proxy-middleware": "^2.0.6",
-    "jquery": "3.7.0",
-    "js-file-download": "0.4.12",
-    "lodash": "4.17.21",
-    "mixin-deep": "2.0.1",
-    "moment": "2.29.4",
-    "noty": "3.2.0-beta",
-    "query-string": "^7.1.3",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-dropzone": "14.2.3",
-    "react-ga": "3.3.1",
-    "react-google-charts": "4.0.1",
-    "react-hyperscript-helpers": "2.0.0",
-    "react-markdown": "8.0.7",
-    "react-material-icon-svg": "3.20.0",
-    "react-modal": "3.16.1",
-    "react-paginating": "1.4.0",
-    "react-protected-mailto": "1.0.3",
-    "react-router-dom": "5.3.0",
-    "react-select": "5.7.4",
-    "react-tooltip": "4.5.1",
-    "stackdriver-errors-js": "0.12.0",
-    "tss-react": "4.8.8",
-    "usa-states": "0.0.6",
-    "uuid": "9.0.0",
-    "voca": "1.4.1"
-  },
-  "scripts": {
-    "analyze": "npm run build; source-map-explorer build/static/js/main.*",
-    "start": "react-scripts start",
-    "genschemas": "./scripts/compile-jsonschema.sh",
-    "build": "react-scripts build",
-    "eject": "react-scripts eject",
-    "cypress:open": "CYPRESS_ADMIN=$(cat cypress/fixtures/duos-automation-admin.json) CYPRESS_CHAIR=$(cat cypress/fixtures/duos-automation-chair.json) CYPRESS_MEMBER=$(cat cypress/fixtures/duos-automation-member.json) CYPRESS_RESEARCHER=$(cat cypress/fixtures/duos-automation-researcher.json) CYPRESS_SIGNING_OFFICIAL=$(cat cypress/fixtures/duos-automation-signing-official.json) cypress open",
-    "cypress:open:component": "cypress open --component",
-    "cypress:run": "CYPRESS_ADMIN=$(cat cypress/fixtures/duos-automation-admin.json) CYPRESS_CHAIR=$(cat cypress/fixtures/duos-automation-chair.json) CYPRESS_MEMBER=$(cat cypress/fixtures/duos-automation-member.json) CYPRESS_RESEARCHER=$(cat cypress/fixtures/duos-automation-researcher.json) CYPRESS_SIGNING_OFFICIAL=$(cat cypress/fixtures/duos-automation-signing-official.json) cypress run",
-    "cypress:run:component": "cypress run --component",
-    "cypress:verify": "cypress verify"
-  },
-  "devDependencies": {
-    "@babel/core": "7.22.9",
-    "@babel/eslint-parser": "7.22.9",
-    "@babel/eslint-plugin": "7.22.5",
-    "@babel/plugin-proposal-class-properties": "7.18.6",
-    "@babel/preset-react": "7.22.5",
-    "@types/react": "18.2.17",
-    "ajv-cli": "5.0.0",
-    "cypress": "12.17.2",
-    "eslint": "7.32.0",
-    "eslint-plugin-flowtype": "7.0.0",
-    "eslint-plugin-react": "7.33.1",
-    "google-auth-library": "9.0.0",
-    "html-webpack-plugin": "5.5.3",
-    "prop-types": "15.8.1",
-    "react-error-overlay": "6.0.11",
-    "react-scripts": "5.0.1",
-    "source-map-explorer": "2.5.3",
-    "typescript": "4.9.5",
-    "webpack": "5.88.2"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ],
-  "overrides": {
-    "ajv-cli": {
-      "fast-json-patch": "^3.1.1"
+    "name": "duos-ui",
+    "version": "1.0.0",
+    "private": true,
+    "dependencies": {
+      "@databiosphere/bard-client": "^0.1.0",
+      "@emotion/react": "^11.11.1",
+      "@emotion/styled": "11.11.0",
+      "@mui/icons-material": "5.14.3",
+      "@mui/material": "5.14.5",
+      "ajv": "8.12.0",
+      "ajv-formats": "2.1.1",
+      "axios": "^1.4.0",
+      "bootstrap": "3.4.1",
+      "dompurify": "3.0.5",
+      "http-proxy-middleware": "^2.0.6",
+      "jquery": "3.7.0",
+      "js-file-download": "0.4.12",
+      "lodash": "4.17.21",
+      "mixin-deep": "2.0.1",
+      "moment": "2.29.4",
+      "noty": "3.2.0-beta",
+      "query-string": "^7.1.3",
+      "react": "18.2.0",
+      "react-dom": "18.2.0",
+      "react-dropzone": "14.2.3",
+      "react-ga": "3.3.1",
+      "react-google-charts": "4.0.1",
+      "react-hyperscript-helpers": "2.0.0",
+      "react-markdown": "8.0.7",
+      "react-material-icon-svg": "3.20.0",
+      "react-modal": "3.16.1",
+      "react-paginating": "1.4.0",
+      "react-protected-mailto": "1.0.3",
+      "react-router-dom": "5.3.0",
+      "react-select": "5.7.4",
+      "react-tooltip": "4.5.1",
+      "stackdriver-errors-js": "0.12.0",
+      "tss-react": "4.9.0",
+      "usa-states": "0.0.6",
+      "uuid": "9.0.0",
+      "voca": "1.4.1"
+    },
+    "scripts": {
+      "analyze": "npm run build; source-map-explorer build/static/js/main.*",
+      "start": "react-scripts start",
+      "genschemas": "./scripts/compile-jsonschema.sh",
+      "build": "react-scripts build",
+      "eject": "react-scripts eject",
+      "cypress:open": "CYPRESS_ADMIN=$(cat cypress/fixtures/duos-automation-admin.json) CYPRESS_CHAIR=$(cat cypress/fixtures/duos-automation-chair.json) CYPRESS_MEMBER=$(cat cypress/fixtures/duos-automation-member.json) CYPRESS_RESEARCHER=$(cat cypress/fixtures/duos-automation-researcher.json) CYPRESS_SIGNING_OFFICIAL=$(cat cypress/fixtures/duos-automation-signing-official.json) cypress open",
+      "cypress:open:component": "cypress open --component",
+      "cypress:run": "CYPRESS_ADMIN=$(cat cypress/fixtures/duos-automation-admin.json) CYPRESS_CHAIR=$(cat cypress/fixtures/duos-automation-chair.json) CYPRESS_MEMBER=$(cat cypress/fixtures/duos-automation-member.json) CYPRESS_RESEARCHER=$(cat cypress/fixtures/duos-automation-researcher.json) CYPRESS_SIGNING_OFFICIAL=$(cat cypress/fixtures/duos-automation-signing-official.json) cypress run",
+      "cypress:run:component": "cypress run --component",
+      "cypress:verify": "cypress verify"
+    },
+    "devDependencies": {
+      "@babel/core": "7.22.10",
+      "@babel/eslint-parser": "7.22.10",
+      "@babel/eslint-plugin": "7.22.10",
+      "@babel/plugin-proposal-class-properties": "7.18.6",
+      "@babel/preset-react": "7.22.5",
+      "@types/react": "18.2.20",
+      "ajv-cli": "5.0.0",
+      "cypress": "12.17.4",
+      "eslint": "7.32.0",
+      "eslint-plugin-flowtype": "7.0.0",
+      "eslint-plugin-react": "7.33.2",
+      "google-auth-library": "9.0.0",
+      "html-webpack-plugin": "5.5.3",
+      "prop-types": "15.8.1",
+      "react-error-overlay": "6.0.11",
+      "react-scripts": "5.0.1",
+      "source-map-explorer": "2.5.3",
+      "typescript": "4.9.5",
+      "webpack": "5.88.2"
+    },
+    "browserslist": [
+      ">0.2%",
+      "not dead",
+      "not ie <= 11",
+      "not op_mini all"
+    ],
+    "overrides": {
+      "ajv-cli": {
+        "fast-json-patch": "^3.1.1"
+      }
     }
   }
-}
